backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-allauth>=0.57.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
argon2-cffi>=23.1.0
Pillow>=10.1.0
